fix(TUP-21175):fix JUnit (#2068)

+++ main/plugins/org.talend.repository.items.importexport/src/main/java/org/talend/repository/items/importexport/handlers/ImportExportHandlersManager.java
@@ -795,7 +795,8 @@ public class ImportExportHandlersManager {
             if (root instanceof File) {
                 final File workingFolder = (File) root;
                 File tmpdir = new File(System.getProperty("java.io.tmpdir")); //$NON-NLS-1$
-                if (workingFolder.toString().startsWith(tmpdir.toString())) { // remove from temp
+                boolean isTest = Boolean.getBoolean("junit.test.import"); //$NON-NLS-1$
+                if (workingFolder.toString().startsWith(tmpdir.toString()) && !isTest) { // remove from temp
                     FilesUtils.deleteFolder(workingFolder, true);
                 }
             }
